requirements.txt
asgiref==3.11.1
dj-database-url==2.3.0
Django==6.0.2
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
psycopg[binary]==3.2.13
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
